renamed StepDescriptionStore
